+++ e2e/wallets/bip85-subaccounts/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @coinmasters/integration-coins
 
+## 1.8.21
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.8.21
+
 ## 1.8.20
 
 ### Patch Changes